refactor: delete cli/common.py, move helpers to their domain packages
managed_mount moved to disks/cli/helpers, preflight helpers already
existed in preflight/cli/helpers — consumers now import from the
canonical locations.
